pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vault-health-mcp"
version = "0.2.0"
description = "MCP server for Obsidian vault structural health checks. Find broken links, orphaned notes, missing frontmatter — and auto-repair them."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Toygar Ermin" },
]
keywords = ["mcp", "obsidian", "health", "lint", "knowledge-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "mcp>=1.20.0",
    "pyyaml>=6.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
]

[project.scripts]
vault-health-mcp = "vault_health.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["vault_health"]
